refactor: remove 'class' from refactored classes
