+++ src/libs/engine/query_interpreter/index.ts
@@ -1,3 +1,3 @@
-import { mountInputDescriptor } from './ohm_interpreter/ohm_interpreter.class';
+import { mountInputDescriptor } from './ohm_interpreter/ohm_interpreter';
 
 export { mountInputDescriptor };
